--- src/router/index.js
@@ -1,34 +1,0 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import Home from '../views/Home.vue'
-
-Vue.use(VueRouter)
-
-const routes = [
-  {
-    path: '/',
-    name: 'Home',
-    component: Home
-  },
-  {
-    path: '/categories',
-    name: 'Categories',
-    component: () => import(/* webpackChunkName: "categories" */ '../views/Categories.vue')
-  },
-  {
-    path: '/preselection',
-    name: 'Preselection',
-    component: () => import(/* webpackChunkName: "preselection" */ '../views/Preselection.vue')
-  },
-  {
-    path: '/about',
-    name: 'About',
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  }
-]
-
-const router = new VueRouter({
-  routes
-})
-
-export default router
